Changed Names of implementations of DataType, added new InputType
