+++ src/main/java/model/job/metadata/InputType.java
@@ -3,6 +3,7 @@ package model.job.metadata;
 public enum InputType {
 	Literal,
 	ComplexData,
-	BoundingBox
+	BoundingBox,
+	URLParameter
 
 }
