+++ src/main/java/model/job/metadata/Service.java
@@ -2,6 +2,12 @@ package model.job.metadata;
 
 import java.util.List;
 
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.annotation.JsonInclude.Include;
+
+@JsonInclude(Include.NON_NULL)
+@JsonIgnoreProperties(ignoreUnknown = true)
 public class Service {
 	private String name;
 	private ResourceMetadata resourceMetadata;
